Feat: access token을 통한 사용자 정보 조회 기능 구현

+++ backend/auth/src/main/java/wontouch/auth/util/jwt/JwtFilter.java
@@ -25,7 +25,7 @@ public class JwtFilter extends OncePerRequestFilter {
         String token = getJwtFromRequest(request);
 
         if (token != null && jwtProvider.validateToken(token)) {
-            Long userId = jwtProvider.getAuthentication(token);
+            int userId = jwtProvider.getAuthentication(token);
             UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userId, null, new ArrayList<>());
             SecurityContextHolder.getContext().setAuthentication(authentication);
         }
